Add prevalence search parameter panel
	- Refactored prevalence search

Ticket: #826
