+++ src/app/prevalence/components/PrevalenceSideContent.tsx
@@ -1,24 +1,18 @@
-import React, { useState } from "react";
-import { Box, Button } from "@mui/material";
 import Search from "@mui/icons-material/Search";
+import { Box, Button } from "@mui/material";
+import React from "react";
 import { useTranslation } from "react-i18next";
 import { FilterMultiSelectionComponent } from "../../evaluations/components/FilterMultiSelectionComponent";
-import { usePrevalenceSideContent } from "./usePrevalenceSideContent";
+import { usePrevalenceFilters } from "./PrevalenceDataContext";
 
 export function PrevalenceSideContent(): JSX.Element {
     const { t } = useTranslation(["PrevalencePage"]);
     const {
-        handleMicroorganismSelectionChange,
-        handleAnimalSpeciesSelectionChange,
+        selectedMicroorganisms,
+        setSelectedMicroorganisms,
         microorganismOptions,
-        animalSpeciesOptions,
-    } = usePrevalenceSideContent();
-
-    // Local state for selections
-    const [localSelectedMicroorganisms, setLocalSelectedMicroorganisms] =
-        useState<string[]>([]);
-    const [localSelectedAnimalSpecies, setLocalSelectedAnimalSpecies] =
-        useState<string[]>([]);
+        callAPI,
+    } = usePrevalenceFilters();
 
     // Convert options to expected format for the FilterMultiSelectionComponent
     const microorganismSelectionOptions = microorganismOptions.map(
@@ -27,27 +21,6 @@ export function PrevalenceSideContent(): JSX.Element {
             displayName: option,
         })
     );
-    const animalSpeciesSelectionOptions = animalSpeciesOptions.map(
-        (option) => ({
-            value: option,
-            displayName: option,
-        })
-    );
-
-    // Function to update local state when selections change
-    const handleLocalMicroorganismsChange = (selected: string[]): void => {
-        setLocalSelectedMicroorganisms(selected);
-    };
-
-    const handleLocalAnimalSpeciesChange = (selected: string[]): void => {
-        setLocalSelectedAnimalSpecies(selected);
-    };
-
-    // Function to call when the search button is clicked
-    const handleSearch = (): void => {
-        handleMicroorganismSelectionChange(localSelectedMicroorganisms);
-        handleAnimalSpeciesSelectionChange(localSelectedAnimalSpecies);
-    };
 
     return (
         <Box
@@ -60,38 +33,24 @@ export function PrevalenceSideContent(): JSX.Element {
             }}
         >
             <FilterMultiSelectionComponent
-                selectedItems={localSelectedMicroorganisms}
+                selectedItems={selectedMicroorganisms}
                 selectionOptions={microorganismSelectionOptions}
                 name="microorganisms"
                 label={t("Microorganism")}
                 actions={{
                     // eslint-disable-next-line @typescript-eslint/no-explicit-any
                     handleChange: (event: any) =>
-                        handleLocalMicroorganismsChange(event.target.value),
+                        setSelectedMicroorganisms(event.target.value),
                 }}
             />
 
-            <FilterMultiSelectionComponent
-                selectedItems={localSelectedAnimalSpecies}
-                selectionOptions={animalSpeciesSelectionOptions}
-                name="animalSpecies"
-                label={t("Animal Species")}
-                // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                actions={{
-                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                    handleChange: (event: any) =>
-                        handleLocalAnimalSpeciesChange(event.target.value),
-                }}
-            />
-
-            {/* Filter Button */}
             <Box
                 sx={{ display: "flex", justifyContent: "center", marginTop: 2 }}
             >
                 <Button
                     variant="contained"
                     startIcon={<Search />}
-                    onClick={handleSearch}
+                    onClick={callAPI}
                 >
                     {t("SEARCH")}
                 </Button>
