+++ src/locales/de/PrevalencePage.json
@@ -6,7 +6,8 @@
   "SEARCH": "Suchen",
   "SELECT_ALL": "Alle auswählen",
   "DESELECT_ALL": " Alle abwählen",
-
+  "ALL_VALUES": "Alle Werte",
+  "MICROORGANISMS": "Mikroorganismen",
   "microorganismOptions": {
     "BAYLISASCARIS_PROCYONIS": "Baylisascaris procyonis",
     "CARBA_E_COLI": "CARBA-E. coli",
